Some polish and minor refactors
